requirements.txt
flask==2.0.1
gunicorn==20.1.0
numpy==1.22.0
tensorflow-cpu==2.8.0
pillow==9.0.0
matplotlib==3.5.1
werkzeug==2.0.2
